Here for reference, I could not get the express proxy middleware to work with next js no matter what I did.
My main aim was one of two possibilities.

Ideally I wanted all routes to go through the reverse proxy and then additionally after that have nextJS render some routes using the response from the proxy. This failed as the proxy middleware insists on handling the response and sending it back to the client, it won't delegate to another middleware in the chain.

So I wanted to know if I could use the proxy for all routes except /a /b and /c/blah, which would be purely rendered by NextJS, however that failed as although you can specify app.render on these routes, all routes need to have `handle` from NextJS called on them to render correctly. I tried adding handle as an extra middleware to those routes only but this also failed.

I'm now going to take a different approach.