Added a NextJS app to act as the stitching layer. It's called PLP, however it does proxy the index too and pretty much any other page I think.
Either way, that's not the point. I've tried to integrate it with OC, however the structure delivered by OC isn't quite what I was expecting and I don't have time to mess with it now. So for now, I'm making services based on Chris' header service.

I have included the OC component for the moment, however it will be removed next commit.

My explanation of the attempt will come with that code.