Logging
Added hybris url env var to index.js
Added header url as it doesn't use the upstream as the upstream docker internal url is https and the component only supports http. Uses the docker-compose networking layer.
Changed upstream url in docker-compose.yml to be the internal network url rather than the docker host url. This requires the docker containers for hybris to be on the network tu.
Also added a few commented out lines to aid local development mounting for the stitcher.